refactor: move src/game-logic/reducers.js to src/game-logic/reducers/index.js
